added python library for job controls Signed-off-by: acolijn <auke.pieter.colijn@gmail.com>
