+++ slowdaq.cc
@@ -26,48 +26,46 @@ int main(int argc, char **argv)
             case 'i': // name of driver
                 DriverFilename = optarg;
                 break;
-	    case 'f':
-		fastOn = true;
-		break;
+            case 'f':
+                fastOn = true;
+                break;
             default:
                 exit(-1);
         }
     }
-      // driver for the daq processing..... provided from the python script
-      //cout << "The driver is: " << DriverFilename << endl;
-      bool slowOn = true;
-      driver* myDriver = new driver(DriverFilename, fastOn, slowOn);
-      // TApplication is needed to plot a canvas with an event
-
-      // create an instance of teh daq datatype: controls all the binary file handling
-      sdaq myDaq(myDriver);
-      // root management
-      srootdriver myRoot(myDriver);
-      ULong64_t new_stime;
-      
-      // loop over the events
-      int totalnumberofevents = myDriver->getNEvent();
-
-	  //cout << "Slow on specified" << endl;
-	  slowevent *sev;
-	  int totalslowevents = myDaq.GetSlowFileSize();
-	  //cout << " I think there are " << totalslowevents << " slow events" << endl;
-	  ULong64_t old_stime = 0;
-	
-	  for (int iSlowEv = 0; iSlowEv < totalslowevents; iSlowEv++){
-	      sev = myDaq.readSlowEvent();
-	      new_stime = myRoot.SlowFill(sev, old_stime);
-	      //cout << "i = " << iSlowEv << endl;
-	      old_stime = new_stime;
-	      delete sev;
-	  }
-	//cout << "I am done processing slow events!!!! " << endl;
-
-
-      //delete theApp;
-      // write run parameters to file
-      //myRoot.writeParameters(myDriver);
-      myRoot.Close();
-      delete myDriver;
-      return 0;
+    // driver for the daq processing..... provided from the python script
+    cout << "MAIN:: slowdaq driver file = " << DriverFilename << endl;
+    bool slowOn = true;
+    driver* myDriver = new driver(DriverFilename, fastOn, slowOn);
+    // TApplication is needed to plot a canvas with an event
+    // create an instance of the daq datatype: controls all the binary file handling
+    sdaq myDaq(myDriver);
+    // root management
+    srootdriver myRoot(myDriver);
+    ULong64_t new_stime;
+    
+    // loop over the events
+    int totalnumberofevents = myDriver->getNEvent();
+    
+    slowevent *sev;
+    int totalslowevents = myDaq.GetSlowFileSize();
+    cout << "MAIN:: Found " << totalslowevents << " slow events" << endl;
+    ULong64_t old_stime = 0;
+    
+    for (int iSlowEv = 0; iSlowEv < totalslowevents; iSlowEv++){
+        if(iSlowEv%1000 == 0) cout << "Processed "<<iSlowEv<<" slow events"<<endl;
+        sev = myDaq.readSlowEvent();
+        new_stime = myRoot.SlowFill(sev, old_stime);
+        old_stime = new_stime;
+        delete sev;
+    }
+    //cout << "I am done processing slow events!!!! " << endl;
+    
+    
+    //delete theApp;
+    // write run parameters to file
+    //myRoot.writeParameters(myDriver);
+    myRoot.Close();
+    delete myDriver;
+    return 0;
 }
